Add support for arbitrarily complex filter expressions
BooleanExpressionParser supports parsing of strings representing arbitrarily complex boolean expressions into predicates.  It is implemented using the shunting yard algorithm, and it supports AND, OR, and NOT compositors, as well as bracketing ("(" and ")").

AND can be specified as "&&" or "&"; OR can be specified as "||" or "|"; NOT is always "!".  These make it intuitive for both programmers and non-programmers.

FilterCommandParser uses BooleanExpressionParser to parse the filter predicate.

Whitespace between adjacent tokens is optional unless the two adjacent tokens can also be interpreted as a single token.

The field key (e.g. "d", "due") must be alphabetic; the test phrase (e.g. "1/10/2018"), if unquoted, can contain any character that is not used by the expression parser ('&', '|', '!', '(', ')') and is not a quote character (''', '"').  Quoted test phrases can contain any character apart from the kind of quote used to begin the test phrase.