Filter: Implement matching with any textual and date field
When the filter field and filter operator (':', '<', '>', '=') are omitted, the test phrase is matched against the name, deadline, and tags fields.  Priority and frequency are not matched as they often cause many tasks to match, and this is usually not intended (and annoying) to the user.  The convenience operator is used for matching.

Tests will come in a later commit.