ClockRule: Migrate to ClockExtension
Currently, we use ClockRule to implement a clock for unit testing.

Rules are part of JUnit 4, and have been replaced by extensions in JUnit 5. As we migrate to JUnit 5, we should update ClockRule to follow the new naming convenion and use JUnit 5 extension style.

Let's rewrite ClockRule to implement an injectable clock using JUnit5 callbacks.